SOLR-16243: Return an implementor/planner after walking the parse tree
