+++ solr/modules/sql/src/java/org/apache/solr/handler/sql/SolrProject.java
@@ -53,7 +53,10 @@ class SolrProject extends Project implements SolrRel {
     return super.computeSelfCost(planner, mq).multiplyBy(0.1);
   }
 
-  public void implement(Implementor implementor) {
+  public Implementor implement(Implementor implementor) {
+
+    implementor.visitChild(0, getInput());
+
     final SolrRules.RexToSolrTranslator translator =
         new SolrRules.RexToSolrTranslator(
             (JavaTypeFactory) getCluster().getTypeFactory(),
@@ -65,6 +68,6 @@ class SolrProject extends Project implements SolrRel {
       implementor.addFieldMapping(name, expr, false);
     }
 
-    implementor.visitChild(0, getInput());
+    return implementor;
   }
 }
